% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpp_cpp_rd.r
\name{bpp_solver_dpp}
\alias{bpp_solver_dpp}
\title{bpp_solver_dpp}
\usage{
bpp_solver_dpp(id, ldhw, m)
}
\arguments{
\item{id}{<vector>

 id order id <integer> vector - should sorted or at least grouped w.r.t order id}

\item{ldhw}{<matrix>

 it order list

 - l, d, h, w it scale along x, y, z and w <numeric>

 it columns should corresponding to id}

\item{m}{<matrix>

 m a bin list

 - l, d, h, w bn scale along x, y, z and w <numeric>

 m should sorted w.r.t preference}
}
\value{
bppSgl
}
\description{
main solver of e-commerce warehouse packing algorithm
}
\details{
bpp init a list of order on sku in data.frame it - oid, sku, l, d, h, w:
  order id oid, stock keeping unit sku, length l, depth d, height h and weight w,

 and also a list of available bn in data.frame bn - id, l, d, h, w:
  bn id, length l, depth d, height h, and weight limit w, sorted by peference often smaller prefered,

 and a single weight limit wlmt applied on all bin.

 bpp solver would solve

  select least number of bn for packing each order w.r.t bn size and weight limit
   and make sure the bn selected are as small as possible.
}
\seealso{
Other bpp_solver_dpp: \code{\link{bpp_solver_dpp_wrapper}},
  \code{\link{bpp_solver_sgl}}
}

