% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bpp_solver_dpp_wrapper}
\alias{bpp_solver_dpp_wrapper}
\title{bpp_solver_dpp_wrapper}
\usage{
bpp_solver_dpp_wrapper(it, bn)
}
\arguments{
\item{it}{<data.frame>

 it order itemSKU list

 - oid: order id <integer>

 - sku: stock keeping unit - it id <character>

 - l, d, h, w it scale along x, y, z and w <numeric>

 - w will be used as constraint while l, d, h will be used as both constraint and objective

 it must be sorted w.r.t oid}

\item{bn}{<data.frame>

 bn a bin list

 - id: bin id <character>

 - l, d, h, w bn scale along x, y, z and w <numeric>

 bn must be sorted w.r.t preference and have l >= d >= h}
}
\value{
sn <list>

 sn solution - it order itemSKU list with tid, bid, and x, y, z <data.frame>

 - oid: order id inherited from it <character>

 - tid: ticket id implied one order can be packed using several ticket id <character>

   each ticket id corresponding to a bid bin id which indicates which bin to use for packing

 - bid: bin id which bn in bn list should be used in pakcing <character>

 - sku: stock keeping unit it id <character>

 - x, y, z it position in the bin <numeric>

 - l, d, h it scale along x, y, z <numeric>

   l, d, h is not inherited from it as it can be rotated to different orientation for packing

 - w it weight scale inherited from it <numeric>
}
\description{
a wrapper over bpp_solver_dpp and expose an nicer r interface
}
\seealso{
Other bpp_solver_dpp: \code{\link{bpp_solver_dpp}},
  \code{\link{bpp_solver_sgl}}
}

