% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpp_cpp_rd.r
\name{bpp_solver_sgl}
\alias{bpp_solver_sgl}
\title{bpp_solver_sgl}
\usage{
bpp_solver_sgl(ldhw, m)
}
\arguments{
\item{ldhw}{<matrix>

 it order list

 - l, d, h, w it scale along x, y, z and w <numeric>

 it columns should corresponding to id}

\item{m}{<matrix>

 m a bin list

 - l, d, h, w bn scale along x, y, z and w <numeric>

 m should sorted w.r.t preference}
}
\value{
bppSgl
}
\description{
subroutine of bpp_solver_dpp
}
\details{
fit a single order into bn list, call gbp4d_solver_dpp_filt() as main solver.
}
\seealso{
Other bpp_solver_dpp: \code{\link{bpp_solver_dpp_wrapper}},
  \code{\link{bpp_solver_dpp}}
}

