% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpp_viewer.r
\name{bpp_viewer}
\alias{bpp_viewer}
\title{bpp_viewer}
\usage{
bpp_viewer(sn, title = NULL, subtitle = NULL)
}
\arguments{
\item{sn}{sn bpp_solution from bpp_solver <list>

 - it item <data.table>

   - oid: order id <integer>

   - sku: stock keeping unit as it id <character>

   - tid: ticket id - an unique id within oid <integer>

   - otid: order id x ticket id - an unique indentifier indicate it with same tid can be packed into one bin <character>

   - bid: bn id <integer>

   - x, y, z it position in bid bin <numeric>

   - l, d, h it scale along x, y, z <numeric>

   - w it weight <numeric>

 - bn bins <data.table>

   - id bn id <character>

   - l bn length limit along x-coordinate <numeric>

   - d bn depth  limit along y-coordinate <numeric>

   - h bn height limit along z-coordinate <numeric>

   - w bn weight limit along w - a separate single dimension <numeric>}

\item{title}{title <character>}

\item{subtitle}{subtitle <character>}
}
\description{
bpp single or multiple order packing solution viewer
}
\seealso{
Other bpp_viewer: \code{\link{bpp_viewer_single}}
}

