% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpp_viewer.r
\name{bpp_viewer_single}
\alias{bpp_viewer_single}
\title{bpp_viewer_single}
\usage{
bpp_viewer_single(it, bn, title = NULL, subtitle = NULL,
  it_rgl_control = NULL, bn_rgl_control = NULL, label_it = TRUE,
  label_bn = TRUE)
}
\arguments{
\item{it}{it item <data.table>

 - id it id <integer>

 - x, y, z it position w.r.t bins <numeric>

 - l, d, h it scale along x, y, z <numeric>

 - w it weight <numeric>

 - auto: cc, wd, txt point and lines color, size, legend <numeric/character, numeric, character>}

\item{bn}{bn bins <data.table>

 - id bn id <integer>

 - l, d, h bn scale <numeric>

 - w bn weight limit <numeric>

 - auto: cc, wd, txt point and lines color, size, legend <numeric/character, numeric, character>}

\item{title}{title <character>}

\item{subtitle}{subtitle <character>}

\item{it_rgl_control}{control the color of it in rgl}

\item{bn_rgl_control}{control the color of bn in rgl}

\item{label_it}{label text on it corner or not}

\item{label_bn}{label text on bn corner or not}
}
\description{
bpp solution viewer on single bin all item
}
\seealso{
Other bpp_viewer: \code{\link{bpp_viewer}}
}

