% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpp_cpp_rd.r
\docType{class}
\name{bppSgl}
\alias{Rcpp_bppSgl}
\alias{Rcpp_bppSgl-class}
\alias{bppSgl}
\title{bppSgl}
\format{An object of class \code{C++Class} of length 1.}
\usage{
bppSgl
}
\description{
bpp solution of a single one order or multiple order
}
\details{
packing it into multiple bn w.r.t bn size and weight limit while select bn as small as possible

 a bppSgl class instance has 6 fields:

  - id: order id <integer>

     list - should sorted or at least grouped w.r.t order id

  - it: it position and scale <matrix>

    - x, y, z, w it position and w in the bin <numeric> (w hold in bn when fit it in bn)

    - l, d, h, w it scale along x, y, z and w <numeric> (w of it itself)

  - bn: bn scale <vector>

    - l, d, h, w bn scale along x, y, z and w <numeric>

  - k: ticket id indicator 0 (if cannot fit into any bin), 1, 2, 3, 4, ... <vector>

  - kb: ticket bn id indicator - which bn to use for packing each ticket <vector>

  - ok: ok a quick indicator of any it can not fit into any bn? <bool>
}
\keyword{datasets}

