% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpp_viewer.r
\name{create_it_cube3d}
\alias{create_it_cube3d}
\title{create_it_cube3d}
\usage{
create_it_cube3d(id, x, y, z, l, d, h, cc, wd, txt, itxt = TRUE)
}
\arguments{
\item{id}{id}

\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{z}{z-coordinate}

\item{l}{length along x-coordinate}

\item{d}{depth  along y-coordinate}

\item{h}{height along z-coordinate}

\item{cc}{color}

\item{wd}{width}

\item{txt}{text}

\item{itxt}{plot text or not}
}
\description{
subroutine of bpp_viewer_single
}
\details{
add it or bn on current rgl device
}

