% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpp.r
\docType{package}
\name{gbp}
\alias{gbp}
\alias{gbp-package}
\title{gbp}
\description{
a collection of 1d, 2d, 3d and 4d bin packing problem solver
}
\section{solver}{


 r-level:

  wrapper over c-level function aims solving e-commerce bin packing problem

   bpp_solver

 c-level:

  core class and solver on 1d, 2d, 3d and 4d bpp

   gbp1d_solver_dpp

   gbp2d_solver_dpp

   gbp2d_solver_dpp_filt

   gbp3d_solver_dpp

   gbp3d_solver_dpp_filt

   gbp4d_solver_dpp

   gbp4d_solver_dpp_filt

   bpp_solver_sgl

   bpp_solver_dpp
}

\section{optimizer}{


 TODO: implementing a bin-shuffing optimizer?
}

\section{viewer}{


 rgl 3d show packing obtained via bpp_solver

  bpp_viewer
}

