% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp1d_cpp_rd.r
\docType{class}
\name{gbp1d}
\alias{Rcpp_gbp1d}
\alias{Rcpp_gbp1d-class}
\alias{gbp1d}
\title{gbp1d}
\format{An object of class \code{C++Class} of length 1.}
\usage{
gbp1d
}
\description{
generalized bin packing problem in 1 dimension, a.k.a knapsack 0-1 problem.
}
\details{
gbp1d init a profit vector p, a weight vector w, and a weight constraint c,
 gbp1d solver would solve

    maximize sum_{j=1}^{n} p_{j} x_{j}

  subject to sum_{j=1}^{n} w_{j} x_{j} leq c
             x_{j} in {0, 1}, j = 1, ...., n

 and instantiate a gbp1d object with a selectin vector x and an objective z.

 gbp1d is implemented as rcpp class, an instantiate can be solved by calling
  gbp1d_solver_dpp(p, w, c) and gbp1d_solver_min(p, w, c)
}
\seealso{
Other gbp1d: \code{\link{gbp1d_solver_dpp}}
}
\keyword{datasets}

