% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp2d_cpp_rd.r
\name{gbp2d_checkr}
\alias{gbp2d_checkr}
\title{gbp2d_checkr}
\usage{
gbp2d_checkr(sn)
}
\arguments{
\item{sn}{<gbp2d>
gbp2d object from gbp2d_solver_dpp() solution.}
}
\value{
okfit? <bool>
}
\description{
auxilium of gbp2d and gbp2d_solver_dpp
}
\details{
check fit solution is valid:
  no conflict between item and bin, and no conflict between each pair of item.
}
\seealso{
Other gbp2d: \code{\link{gbp2d_solver_dpp}},
  \code{\link{gbp2d}}
}

