% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp2d_cpp_rd.r
\name{gbp2d_it_create_ktlist}
\alias{gbp2d_it_create_ktlist}
\title{gbp2d_it_create_ktlist}
\usage{
gbp2d_it_create_ktlist(bn, it, xp, ktinit, nlmt)
}
\arguments{
\item{bn}{bn scale <vector>
- l, d bn scale along x and y <numeric>}

\item{it}{it position and scale <matrix>
- x, y it position in the bin <numeric>
- l, d it scale along x and y <numeric>}

\item{xp}{xp extreme point position and residual space scale <matrix>
- x, y xp position in the bin <numeric>
- l, d xp residual space scale along x and y <numeric>}

\item{ktinit}{kt candidate scale without position <matrix>
- l, d kt scale along x and y which open to orientation <numeric>}

\item{nlmt}{nlmt: limit on ktlist n max-value}
}
\value{
Ktlist2d
}
\description{
create ktlist from itlist
}
\details{
core function in gbp2d_solver_dpp
  select highest profitable it not yet fit into bn and return all possbile fit w.r.t xp and orientation
}
\note{
should make sure it kt can be fit in bin outside

internal function use in gbp2d_solver_dpp() for creating Ktlist2d object for fit.
}
\seealso{
Other gbp2d_it: \code{\link{Ktlist2d}}
}

