% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp2d_cpp_rd.r
\name{gbp2d_solver_dpp}
\alias{gbp2d_solver_dpp}
\title{gbp2d_solver_dpp}
\usage{
gbp2d_solver_dpp(p, ld, m)
}
\arguments{
\item{p}{p profit of it fit into bn <vector>
- cluster max(l, d) and min(l, d) via gbp2d_solver_dpp_prep_create_p()}

\item{ld}{it position and scale <matrix>
- l, d it scale along x and y, subject to orientation rotation <numeric>}

\item{m}{bn scale <vector>
- l, d bn scale along x and y <numeric>}
}
\value{
gbp2d
 a gbp2d instantiate with p profit, it item (x, y, l, d) position scale matrix, bn bin (l, d) scale vector,
  k selection, o objective, and ok an indicator of all fit or not.
}
\description{
solve gbp2d via extreme point heuristic and best information score fit strategy.
}
\details{
gbp2d init a profit vector p, a length vector l, a depth vector d,
  a length constraint ml, and a depth constraint md on l x d rectangle
  with geometry intepretation.

 gbp2d solver would solve

   maximize   sum_{j=1}^{n} p_{j} k_{j}

   subject to fit (l_{j}, d_{j}) at coordinate (x_{j}, y_{j})
              such that no overlap in ml x md, j = 1, ...., n

 and instantiate a gbp2d object with a x-axis coordinate vector x, a y-axis coordinate vector y,
  a selection vector k, and an objective o.
}
\seealso{
Other gbp2d: \code{\link{gbp2d_checkr}},
  \code{\link{gbp2d}}
}

