% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gbp2d_solver_dpp_prep_create_p}
\alias{gbp2d_solver_dpp_prep_create_p}
\title{gbp2d_solver_dpp_prep_create_p}
\usage{
gbp2d_solver_dpp_prep_create_p(ld, m)
}
\arguments{
\item{ld}{2xN matrix of l, d of it}

\item{m}{2x1 vector of l, d of bn}
}
\value{
p
}
\description{
auxilium of gbp2d_solver_dpp
}
\details{
create p via ld and m via cluster max(l, d) and min(l, d) strategy
}

