% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp2d_viewer.r
\name{gbp2d_viewer}
\alias{gbp2d_viewer}
\title{gbp2d_viewer}
\usage{
gbp2d_viewer(sn, title = NULL, subtitle = NULL)
}
\arguments{
\item{sn}{sn gbp2d object, solution from gbp2d_solver_dpp, see gbp2d.}

\item{title}{title <character>}

\item{subtitle}{subtitle <character>}
}
\description{
gbp2d solution viewer
}

