% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp2d_cpp_rd.r
\name{gbp2q_checkr}
\alias{gbp2q_checkr}
\title{gbp2q_checkr}
\usage{
gbp2q_checkr(sn)
}
\arguments{
\item{sn}{<gbp2q>
gbp2q object from gbp2d_solver_dpp_filt() solution.}
}
\value{
okfit? <bool>
}
\description{
auxilium of gbp2q and gbp2d_solver_dpp_filt
}
\details{
check fit solution is valid:
  no conflict between item and bin, and no conflict between each pair of item.
}
\seealso{
Other gbp2q: \code{\link{gbp2d_solver_dpp_filt}},
  \code{\link{gbp2q}}
}

