% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp2d_viewer.r
\name{gbp2q_viewer}
\alias{gbp2q_viewer}
\title{gbp2q_viewer}
\usage{
gbp2q_viewer(sn, title = NULL, subtitle = NULL)
}
\arguments{
\item{sn}{sn gbp2q object, solution from gbp2d_solver_dpp_filt, see gbp2q.}

\item{title}{title <character>}

\item{subtitle}{subtitle <character>}
}
\description{
gbp2q solution viewer
}

