% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp3d_cpp_rd.r
\name{gbp3d_checkr}
\alias{gbp3d_checkr}
\title{gbp3d_checkr}
\usage{
gbp3d_checkr(sn)
}
\arguments{
\item{sn}{<gbp3d>
gbp3d object from gbp3d_solver_dpp() solution.}
}
\value{
okfit? <bool>
}
\description{
auxilium of gbp3d_solver_dpp
}
\details{
check fit solution is valid:
  no conflict between item and bin, and no conflict between each pair of item.
}
\seealso{
Other gbp3d: \code{\link{gbp3d_solver_dpp}},
  \code{\link{gbp3d}}
}

