% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp3d_cpp_rd.r
\name{gbp3d_it_create_ktlist}
\alias{gbp3d_it_create_ktlist}
\title{gbp3d_it_create_ktlist}
\usage{
gbp3d_it_create_ktlist(bn, it, xp, ktinit, nlmt)
}
\arguments{
\item{bn}{bn scale <vector>
- l, d, h bn scale along x, y, z <numeric>}

\item{it}{it position and scale <matrix>
- x, y, z it position in the bin <numeric>
- l, d, h it scale along x, y, z <numeric>}

\item{xp}{xp extreme point position and residual space scale <matrix>
- x, y, z xp position in the bin <numeric>
- l, d, h xp residual space scale along x, y, z <numeric>}

\item{ktinit}{kt candidate scale without position <matrix>
- l, d, h kt scale along x, y, z which open to orientation <numeric>}

\item{nlmt}{nlmt: limit on ktlist n max-value}
}
\value{
Ktlist3d
}
\description{
create ktlist from itlist
}
\details{
core function in gbp3d_solver_dpp
  select highest profitable it not yet fit into bn and return all possbile fit w.r.t xp and orientation
}
\note{
should make sure it kt can be fit in bin outside

internal function use in gbp3d_solver_dpp() for creating Ktlist3d object for fit.
}
\seealso{
Other gbp3d_it: \code{\link{Ktlist3d}}
}

