% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp3d_cpp_rd.r
\name{gbp3d_solver_dpp}
\alias{gbp3d_solver_dpp}
\title{gbp3d_solver_dpp}
\usage{
gbp3d_solver_dpp(p, ldh, m)
}
\arguments{
\item{p}{p profit of it fit into bn <vector>
- cluster max(l, d) and min(l, d) via gbp3d_solver_dpp_prep_create_p()}

\item{ldh}{it position and scale <matrix>
- l, d, h it scale along x, y, z, subject to orientation rotation <numeric>}

\item{m}{bn scale <vector>
- l, d, h bn scale along x, y, z <numeric>}
}
\value{
gbp3d
 a gbp3d instantiate with p profit, it item (x, y, z, l, d, h) position scale matrix, bn bin (l, d, h) scale vector,
  k selection, o objective, and ok an indicator of all fit or not.
}
\description{
solve gbp3d via extreme point heuristic and best information score fit strategy.
}
\details{
gbp3d init a profit vector p, a length vector l, a depth vector d, a height vector h, and also
  a length constraint ml, a depth constraint md, and a height constraint mh on l x d x h cuboid
  with geometry intepretation.

 gbp3d solver would solve

   maximize   sum_{j=1}^{n} p_{j} k_{j}

   subject to fit (l_{j}, d_{j}, h_{j}) at coordinate (x_{j}, y_{j}, z_{j})
              such that no overlap in ml x md x mh cuboid, j = 1, ......, n

 and instantiate a gbp3d object with a x-axis coordinate vector x, a y-axis coordinate vector y,
  a z-axis coordinate vector z, a selection vector k, and an objective o.
}
\seealso{
Other gbp3d: \code{\link{gbp3d_checkr}},
  \code{\link{gbp3d}}
}

