% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp3d_viewer.r
\name{gbp3d_viewer}
\alias{gbp3d_viewer}
\title{gbp3d_viewer}
\usage{
gbp3d_viewer(sn, title = NULL, subtitle = NULL)
}
\arguments{
\item{sn}{sn gbp3d object, solution from gbp3d_solver_dpp, see gbp3d.}

\item{title}{title <character>}

\item{subtitle}{subtitle <character>}
}
\description{
gbp3d solution viewer
}

