% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp3d_viewer.r
\name{gbp3q_viewer}
\alias{gbp3q_viewer}
\title{gbp3q_viewer}
\usage{
gbp3q_viewer(sn, title = NULL, subtitle = NULL)
}
\arguments{
\item{sn}{sn gbp3q object, solution from gbp3d_solver_dpp_filt, see gbp3q.}

\item{title}{title <character>}

\item{subtitle}{subtitle <character>}
}
\description{
gbp3q solution viewer
}

