% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp4d_cpp_rd.r
\name{gbp4d_checkr}
\alias{gbp4d_checkr}
\title{gbp4d_checkr}
\usage{
gbp4d_checkr(sn)
}
\arguments{
\item{sn}{<gbp4d>
gbp4d object from gbp4d_solver_dpp() solution.}
}
\value{
okfit? <bool>
}
\description{
auxilium of gbp4d_solver_dpp
}
\details{
check fit solution is valid:
  no conflict between item and bin, and no conflict between each pair of item, and no conflict on weight limit.
}
\seealso{
Other gbp4d: \code{\link{gbp4d_solver_dpp}},
  \code{\link{gbp4d}}
}

