% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp4d_cpp_rd.r
\name{gbp4d_it_create_ktlist}
\alias{gbp4d_it_create_ktlist}
\title{gbp4d_it_create_ktlist}
\usage{
gbp4d_it_create_ktlist(bn, it, xp, ktinit, nlmt)
}
\arguments{
\item{bn}{bn scale <vector>
- l, d, h, w bn scale along x, y, z and w <numeric>}

\item{it}{it position and scale <matrix>
- x, y, z, w it position and w in the bin <numeric>
- l, d, h, w it scale along x, y, z and w <numeric>}

\item{xp}{xp extreme point position and residual space scale <matrix>
- x, y, z, w xp position and w in the bin <numeric>
- l, d, h, w xp residual space scale along x, y, z and w <numeric>}

\item{ktinit}{kt candidate scale without position <matrix>
- l, d, h, w kt scale along x, y, z, w which open to orientation <numeric>}

\item{nlmt}{nlmt: limit on ktlist n max-value}
}
\value{
Ktlist4d
}
\description{
create ktlist from itlist
}
\details{
core function in gbp4d_solver_dpp
  select highest profitable it not yet fit into bn and return all possbile fit w.r.t xp and orientation
}
\note{
should make sure it kt can be fit in bin outside

internal function use in gbp4d_solver_dpp() for creating Ktlist4d object for fit.
}
\seealso{
Other gbp4d_it: \code{\link{Ktlist4d}}
}

