% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gbp4d_solver_dpp_prep_create_p}
\alias{gbp4d_solver_dpp_prep_create_p}
\title{gbp4d_solver_dpp_prep_create_p}
\usage{
gbp4d_solver_dpp_prep_create_p(ldhw, m)
}
\arguments{
\item{ldhw}{4xN matrix of l, d, h, w of it}

\item{m}{4x1 vector of l, d, h, w of bn}
}
\value{
p
}
\description{
auxilium of gbp4d_solver_dpp
}
\details{
create p via ldhw and m via cluster w, cluster max(l, d, h) and area strategy
}

