% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp4d_viewer.r
\name{gbp4d_viewer}
\alias{gbp4d_viewer}
\title{gbp4d_viewer}
\usage{
gbp4d_viewer(sn, title = NULL, subtitle = NULL)
}
\arguments{
\item{sn}{sn gbp4d object, solution from gbp4d_solver_dpp, see gbp4d.}

\item{title}{title <character>}

\item{subtitle}{subtitle <character>}
}
\description{
gbp4d solution viewer
}

