% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp4d_cpp_rd.r
\name{gbp4q_checkr}
\alias{gbp4q_checkr}
\title{gbp4q_checkr}
\usage{
gbp4q_checkr(sn)
}
\arguments{
\item{sn}{<gbp4q>
gbp4q object from gbp4d_solver_dpp_filt() solution.}
}
\value{
okfit? <bool>
}
\description{
auxilium of gbp4d_solver_dpp_filt
}
\details{
check fit solution is valid:
  no conflict between item and bin, and no conflict between each pair of item, and no conflict on weight limit.
}
\seealso{
Other gbp4q: \code{\link{gbp4d_solver_dpp_filt}},
  \code{\link{gbp4q}}
}

