% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp4d_viewer.r
\name{gbp4q_viewer}
\alias{gbp4q_viewer}
\title{gbp4q_viewer}
\usage{
gbp4q_viewer(sn, title = NULL, subtitle = NULL)
}
\arguments{
\item{sn}{sn gbp4q object, solution from gbp4d_solver_dpp_filt, see gbp4q.}

\item{title}{title <character>}

\item{subtitle}{subtitle <character>}
}
\description{
gbp4q solution viewer
}

