% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp2d_cpp_rd.r
\docType{class}
\name{Ktlist2d}
\alias{Ktlist2d}
\alias{Rcpp_Ktlist2d}
\alias{Rcpp_Ktlist2d-class}
\title{Ktlist2d}
\format{An object of class \code{C++Class} of length 1.}
\usage{
Ktlist2d
}
\description{
Ktlist2d hold multiple kt for recursive fit
}
\details{
Ktlist2d hold multiple kt via consider all possible fit onto different xp and different rotation and nlimit

 a Ktlist2d class instance has 4 fields:

  - n: length of kt candidate position scale vector list

  - kt: candidate (x, y, l, d) fit of it current investigating

  - xp: candidate extreme point list after kt fit into each corresponding (x, y, l, d) position scale

  - s: score of each kt fit: calculate overall extrem point residual space entropy as score, the smaller the better,
  since smaller entropy indicate concentrated residual space and less number of extreme point.
}
\note{
internal cpp class use in gbp2d_solver_dpp()
}
\seealso{
Other gbp2d_it: \code{\link{gbp2d_it_create_ktlist}}
}
\keyword{datasets}

