% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp3d_cpp_rd.r
\docType{class}
\name{Ktlist3d}
\alias{Ktlist3d}
\alias{Rcpp_Ktlist3d}
\alias{Rcpp_Ktlist3d-class}
\title{Ktlist3d}
\format{An object of class \code{C++Class} of length 1.}
\usage{
Ktlist3d
}
\description{
Ktlist3d hold multiple kt for recursive fit
}
\details{
Ktlist3d hold multiple kt via consider all possible fit onto different xp and different rotation and nlimit

 a Ktlist3d class instance has 4 fields:

  - n: length of kt candidate position scale vector list

  - kt: candidate (x, y, z, l, d, h) fit of it current investigating

  - xp: candidate extreme point list after kt fit into each corresponding (x, y, z, l, d, h) position scale

  - s: score of each kt fit: calculate overall extrem point residual space entropy as score, the smaller the better,
  since smaller entropy indicate concentrated residual space and less number of extreme point.
}
\note{
internal cpp class use in gbp3d_solver_dpp()
}
\seealso{
Other gbp3d_it: \code{\link{gbp3d_it_create_ktlist}}
}
\keyword{datasets}

