% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp4d_cpp_rd.r
\docType{class}
\name{Ktlist4d}
\alias{Ktlist4d}
\alias{Rcpp_Ktlist4d}
\alias{Rcpp_Ktlist4d-class}
\title{Ktlist4d}
\format{An object of class \code{C++Class} of length 1.}
\usage{
Ktlist4d
}
\description{
Ktlist4d hold multiple kt for recursive fit
}
\details{
Ktlist4d hold multiple kt via consider all possible fit onto different xp and different rotation and nlimit

 a Ktlist4d class instance has 4 fields

  - n: length of kt candidate position scale vector list

  - kt: candidate (x, y, z, w, l, d, h, w) fit of it current investigating

        x, y, z, w - weight holding in bn when fit; l, d, h, w - weight of it itself

  - xp: candidate extreme point list after kt fit into each corresponding (x, y, z, l, d, h) position scale

  - s: score of each kt fit: calculate overall extrem point residual space entropy as score, the smaller the better,
  since smaller entropy indicate concentrated residual space and less number of extreme point.
}
\note{
internal cpp class use in gbp4d_solver_dpp()
}
\seealso{
Other gbp4d_it: \code{\link{gbp4d_it_create_ktlist}}
}
\keyword{datasets}

